renamed h to hpp files

+++ Led.cpp
@@ -1,4 +1,4 @@
-#include "Led.h"
+#include "Led.hpp"
 
 Led::Led(int pin) :
 _pin (pin),
@@ -17,7 +17,7 @@ void Led::setDuty(int v) {
     analogWrite(_pin, v, SWITCH_FREQ);
 }
 
-void Led::strobeFreq(float freq) {
+void Led::strobeFreq(double freq) {
     freq = constrain(freq, 0, 65535);
     int p = 1000.0 / freq;
     _timer.changePeriod(p);
